Now using radius as field input instead of inverse stddev thing
Basically, the previous method was hacky and there wasn't much
understanding. Still having less understanding than I should,
I ran the numbers and found that the max radius produced a power series.
We can aproximate the max radius using 4.761*sqrt(x). Since requiring
the designer to give x is a bit unruly, I switched it to take in the
radius y and calculate the "stdev" in the equation using stdev = (radius/4.761)^2.

It seems to work pretty well!